lazy load npgsql data source to resolve integration tests connection string issue
