+++ Apps/Common/src/AspNetConfiguration/Modules/Db.cs
@@ -42,15 +42,19 @@ namespace HealthGateway.Common.AspNetConfiguration.Modules
             bool isSensitiveDataLoggingEnabled = section.GetValue("Enabled", false);
             logger.LogDebug("Sensitive Data Logging is enabled: {IsSensitiveDataLoggingEnabled}", isSensitiveDataLoggingEnabled);
 
-            NpgsqlDataSourceBuilder dataSourceBuilder = new(configuration.GetConnectionString("GatewayConnection"));
-            dataSourceBuilder.EnableDynamicJson();
-            NpgsqlDataSource dataSource = dataSourceBuilder.Build();
+            services.AddSingleton(_ =>
+            {
+                NpgsqlDataSourceBuilder dataSourceBuilder = new(configuration.GetConnectionString("GatewayConnection"));
+                dataSourceBuilder.EnableDynamicJson();
+                return dataSourceBuilder.Build();
+            });
 
             services.AddDbContextPool<GatewayDbContext>(
-                options =>
+                (sp, options) =>
                 {
+                    var ds = sp.GetRequiredService<NpgsqlDataSource>();
                     options.UseNpgsql(
-                        dataSource,
+                        ds,
                         x => x.MigrationsHistoryTable("__EFMigrationsHistory", "gateway"));
                     if (isSensitiveDataLoggingEnabled)
                     {
